Add frontend folder, redux and react-router

+++ backend/controllers/goalController.js
@@ -35,6 +35,19 @@ const updateGoal = asyncHandler(async(req, res) => {
         res.status(400)
         throw new Error("Goal not found")
     }
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
     res.status(200).json(updatedGoal)
 })
@@ -48,6 +61,20 @@ const deleteGoal = asyncHandler(async(req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+    
     await goal.deleteOne()
     res.status(200).json({ id: req.params.id})
 })
